src/main/java/uk/co/trinitylogic/leetcode/soultions/SolutionBalancedBinaryTree110.java
package uk.co.trinitylogic.leetcode.soultions;

import uk.co.trinitylogic.leetcode.soultions.models.TreeNode;

public class SolutionBalancedBinaryTree110 {

    // https://leetcode.com/problems/balanced-binary-tree/

    public boolean isBalanced(TreeNode root) {
        if (root == null) return true;
        return helper(root) != -1;
    }

    private int helper(TreeNode node) {
        if (node == null) {
            return 0;
        }
        int l = helper(node.left);
        int r = helper(node.right);
        if (l == -1 || r == -1) return -1;
        if (Math.abs(l - r) > 1) return -1;
        return Math.max(l, r) + 1;
    }
}
